+++ annofabapi/api2.py
@@ -5,9 +5,8 @@ import requests
 from requests.cookies import RequestsCookieJar
 
 import annofabapi.utils
-from annofabapi.api import AnnofabApi
+from annofabapi.api import AnnofabApi, _log_error_response, _raise_for_status
 from annofabapi.generated_api2 import AbstractAnnofabApi2
-from annofabapi.utils import _log_error_response, _raise_for_status
 
 logger = logging.getLogger(__name__)
